fix(client): resolve use_build_context_synchronously linting errors.

+++ client/lib/features/events/features/event_page/presentation/widgets/event_page_meeting_agenda.dart
@@ -223,7 +223,7 @@ class _EventPageMeetingAgendaState extends State<EventPageMeetingAgenda>
       textAlign: TextAlign.start,
     ).show(context: context);
 
-    if (delete) {
+    if (delete && mounted) {
       await alertOnError(context, () => _presenter.deleteAgendaItems());
     }
   }
